Split out libraries from util.h
