+++ base/connection_info.cc
@@ -1,14 +1,15 @@
 /*
  * Copyright (c) 2014 Juniper Networks, Inc. All rights reserved.
  */
+#include "base/connection_info.h"
 
 #include <boost/system/error_code.hpp>
 #include <boost/asio/ip/tcp.hpp>
 #include <boost/asio/ip/address.hpp>
 #include <boost/tuple/tuple_comparison.hpp>
 
-#include <base/connection_info.h>
-#include <base/sandesh/process_info_constants.h>
+#include "base/string_util.h"
+#include "base/sandesh/process_info_constants.h"
 
 namespace process {
