+++ base/lifetime.cc
@@ -5,6 +5,8 @@
 #include "base/lifetime.h"
 
 #include <boost/bind.hpp>
+#include "base/backtrace.h"
+#include "base/time_util.h"
 
 LifetimeRefBase::LifetimeRefBase(LifetimeActor *actor)
         : ref_(this, actor) {
